+++ src/main/java/io/redstudioragnarok/witherconfig/WitherConfig.java
@@ -1,0 +1,22 @@
+package io.redstudioragnarok.witherconfig;
+
+import net.minecraftforge.fml.common.Mod;
+
+import static io.redstudioragnarok.witherconfig.utils.ModReference.ID;
+import static io.redstudioragnarok.witherconfig.utils.ModReference.NAME;
+import static io.redstudioragnarok.witherconfig.utils.ModReference.VERSION;
+
+//   /$$      /$$ /$$   /$$     /$$                                  /$$$$$$                       /$$$$$$  /$$
+//  | $$  /$ | $$|__/  | $$    | $$                                 /$$__  $$                     /$$__  $$|__/
+//  | $$ /$$$| $$ /$$ /$$$$$$  | $$$$$$$   /$$$$$$   /$$$$$$       | $$  \__/  /$$$$$$  /$$$$$$$ | $$  \__/ /$$  /$$$$$$
+//  | $$/$$ $$ $$| $$|_  $$_/  | $$__  $$ /$$__  $$ /$$__  $$      | $$       /$$__  $$| $$__  $$| $$$$    | $$ /$$__  $$
+//  | $$$$_  $$$$| $$  | $$    | $$  \ $$| $$$$$$$$| $$  \__/      | $$      | $$  \ $$| $$  \ $$| $$_/    | $$| $$  \ $$
+//  | $$$/ \  $$$| $$  | $$ /$$| $$  | $$| $$_____/| $$            | $$    $$| $$  | $$| $$  | $$| $$      | $$| $$  | $$
+//  | $$/   \  $$| $$  |  $$$$/| $$  | $$|  $$$$$$$| $$            |  $$$$$$/|  $$$$$$/| $$  | $$| $$      | $$|  $$$$$$$
+//  |__/     \__/|__/   \___/  |__/  |__/ \_______/|__/             \______/  \______/ |__/  |__/|__/      |__/ \____  $$
+//                                                                                                              /$$  \ $$
+//                                                                                                             |  $$$$$$/
+//                                                                                                              \______/
+@Mod(modid = ID, name = NAME, version = VERSION/*, updateJSON = "https://raw.githubusercontent.com/Red-Studio-Ragnarok/witherconfig/main/update.json"*/)
+public class WitherConfig {
+}