+++ src/main/java/io/redstudioragnarok/witherconfig/asm/WitherConfigPlugin.java
@@ -1,4 +1,4 @@
-package io.redstudioragnarok.Replace.asm;
+package io.redstudioragnarok.witherconfig.asm;
 
 import net.minecraftforge.fml.relauncher.IFMLLoadingPlugin;
 import zone.rong.mixinbooter.IEarlyMixinLoader;
@@ -9,8 +9,8 @@ import java.util.Map;
 
 @IFMLLoadingPlugin.Name()
 @IFMLLoadingPlugin.MCVersion("1.12.2")
-@IFMLLoadingPlugin.TransformerExclusions("io.redstudioragnarok.Replace.asm")
-public class ReplacePlugin implements IFMLLoadingPlugin, IEarlyMixinLoader {
+@IFMLLoadingPlugin.TransformerExclusions("io.redstudioragnarok.witherconfig.asm")
+public class WitherConfigPlugin implements IFMLLoadingPlugin, IEarlyMixinLoader {
 
     @Override
     public String[] getASMTransformerClass() {
@@ -38,6 +38,6 @@ public class ReplacePlugin implements IFMLLoadingPlugin, IEarlyMixinLoader {
 
     @Override
     public List<String> getMixinConfigs() {
-        return Arrays.asList("mixins.Replace.json");
+        return Arrays.asList("mixins.witherconfig.json");
     }
 }