+++ src/main/java/io/redstudioragnarok/witherconfig/utils/ModReference.java
@@ -1,0 +1,18 @@
+package io.redstudioragnarok.witherconfig.utils;
+
+import io.redstudioragnarok.witherconfig.Tags;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+/**
+ * This class defines constants for Wither Config.
+ * <p>
+ * They are automatically updated by RFG on compile time, except for the name as Gradle would remove spaces.
+ */
+public class ModReference {
+
+    public static final String ID = Tags.ID;
+    public static final String NAME = "Wither Config";
+    public static final String VERSION = Tags.VERSION;
+    public static final Logger LOG = LogManager.getLogger(NAME);
+}