+++ Changelog.md
@@ -4,4 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project follows to [Ragnarök Versioning Convention](https://shor.cz/ragnarok_versioning_convention).
 
-## [Unreleased] Wither Config Version 0.1 Changelog
+## [Unreleased] Wither Config Version 1 Changelog
+
+### Added
+
+- Configuration for:
+  - The Wither health
+  - The Wither movement speed
+  - The Wither follow range
+  - The Wither armor
+  - The Wither follow distance
+  - The Wither summoning sequence length
+  - The Wither summoning sequence end explosion strength
+  - The Wither skulls explosion strength